starts working on best responses
For now the functions that have been implemented are the following:

- stationary state for a repeated game of two memory-one strategies with discount factor (analytically)
- stationary state for an infinitely repeated game

The stationary state with the discount factor is the function we want
to use in the futute. However, I implemented the other function
so we can check evaluate our methodology by reproducing the
work of Kim, Choi and Baek.